requirements.txt
jupyterlab==4.0.9
numpy==1.26.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.3
pyspark==3.5.0
tqdm==4.66.1
python-dateutil==2.8.2
joblib==1.3.2